session-management-redis-python/requirements.txt
fastapi
uvicorn[standard]
aioredis
py-dotenv